Simplified lint command barreling 🛢️
